DAL/requirements.txt
pika==1.3.2
pyyaml==6.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6